Move api enums to dedicated files
