+++ lib/open_meteo.dart
@@ -7,7 +7,13 @@ export 'src/apis/geocoding.dart';
 export 'src/apis/historical.dart';
 export 'src/apis/marine.dart';
 export 'src/apis/weather.dart';
-
-export 'src/options.dart';
+export 'src/enums/air_quality.dart';
+export 'src/enums/climate.dart';
+export 'src/enums/ensemble.dart';
+export 'src/enums/flood.dart';
+export 'src/enums/historical.dart';
+export 'src/enums/marine.dart';
+export 'src/enums/weather.dart';
 export 'src/exceptions.dart';
+export 'src/options.dart';
 export 'src/response.dart';
