+++ lib/src/apis/flood.dart
@@ -1,7 +1,7 @@
 import '../api.dart';
+import '../enums/flood.dart';
 import '../options.dart';
 import '../response.dart';
-import '../weather_api_openmeteo_sdk_generated.dart';
 
 /// Simulated river discharge at 5 km resolution from 1984 up to 7 months forecast.
 ///
@@ -98,45 +98,4 @@ class FloodApi extends BaseApi {
         'timeformat': 'unixtime',
         'timezone': 'auto',
       };
-}
-
-enum FloodDaily with WeatherParameter<FloodApi, Daily> {
-  river_discharge(Variable.river_discharge),
-  river_discharge_mean(
-    Variable.river_discharge,
-    aggregation: Aggregation.mean,
-  ),
-  river_discharge_median(
-    Variable.river_discharge,
-    aggregation: Aggregation.median,
-  ),
-  river_discharge_max(
-    Variable.river_discharge,
-    aggregation: Aggregation.maximum,
-  ),
-  river_discharge_min(
-    Variable.river_discharge,
-    aggregation: Aggregation.minimum,
-  ),
-  river_discharge_p25(
-    Variable.river_discharge,
-    aggregation: Aggregation.p25,
-  ),
-  river_discharge_p75(
-    Variable.river_discharge,
-    aggregation: Aggregation.p75,
-  );
-
-  @override
-  final Variable variable;
-
-  @override
-  final Aggregation aggregation;
-
-  const FloodDaily(
-    this.variable, {
-    this.aggregation = Aggregation.none,
-  });
-
-  static final Map<int, FloodDaily> hashes = makeHashes(FloodDaily.values);
 }
